Adjust time string formatting
I'm halfway following the guidelines at https://brand.wsu.edu/editorial/style/.
Opting not to use periods in am/pm because it adds a little too much noise in
river and archive views, where they're already displayed in all caps. And not
using them in individual views for the sake of consistency.